Create HomeObj and refactor Home commands to use it for patient-finding
