+++ src/main/java/duke/command/home/HomeOpenSpec.java
@@ -2,8 +2,6 @@ package duke.command.home;
 
 import duke.DukeCore;
 import duke.command.ArgLevel;
-import duke.command.CommandUtils;
-import duke.command.ObjSpec;
 import duke.command.Switch;
 import duke.data.DukeObject;
 import duke.data.Impression;
@@ -11,7 +9,7 @@ import duke.data.Patient;
 import duke.exception.DukeException;
 import duke.ui.context.Context;
 
-public class HomeOpenSpec extends ObjSpec {
+public class HomeOpenSpec extends HomeObjSpec {
     private static final HomeOpenSpec spec = new HomeOpenSpec();
 
     public static HomeOpenSpec getSpec() {
@@ -24,18 +22,6 @@ public class HomeOpenSpec extends ObjSpec {
                 new Switch("bed", String.class, true, ArgLevel.REQUIRED, "b"),
                 new Switch("impression", String.class, true, ArgLevel.NONE, "im")
         );
-    }
-
-    @Override
-    protected void execute(DukeCore core) throws DukeException {
-        super.execute(core);
-        String bed = cmd.getSwitchVal("bed");
-        Patient patient = CommandUtils.findFromHome(core, bed, cmd.getArg());
-        if (patient == null) {
-            core.search(core.patientData.findPatients(cmd.getArg()), cmd);
-        } else {
-            executeWithObj(core, patient);
-        }
     }
 
     @Override
